0) Contributed by Hui-Chuan Lin, some trivial fixes
M      var/README.radiance
M      var/da/da_setup_structures/da_setup_be_regional.inc
M      var/da/da_varbc/da_varbc_update.inc
M      var/da/da_radiance/da_setup_satcv.inc
M      var/run/README
M      Registry/Registry.wrfvar

1) Fix bugs introduced by previous commits:

M      var/build/da.make
D      var/da/da_recursive_filter/da_mat.f90
A  +   var/da/da_recursive_filter/da_rf_cv3.f90
A  +   var/da/da_recursive_filter/da_mat_cv3.f90
D      var/da/da_recursive_filter/da_rfz.f90
D      var/da/da_recursive_filter/da_rf.f90
A  +   var/da/da_recursive_filter/da_rfz_cv3.f90
M      var/da/da_recursive_filter/da_recursive_filter.f90
M      var/da/da_profiler/da_get_innov_vector_profiler.inc
M      var/da/da_setup_structures/da_setup_be_ncep_gfs.inc
M      var/da/da_setup_structures/da_setup_structures.f90

2) Add link option -lpng to match recently upgrade of NCARG on some  
NCAR machine

M      var/obsproc/MAP_plot/Map.csh
M      var/obsproc/MAP_plot/configure.user.ibm


3) Establish some sample case directories under var/test for community  
users and regression tests.

A      var/test/4dvar
A      var/test/4dvar/tl
A      var/test/4dvar/tl/VEGPARM.TBL
A      var/test/4dvar/tl/wrfbdy_d01
A      var/test/4dvar/tl/wrfinput_d01
A      var/test/4dvar/tl/namelist.input.parallel
A      var/test/4dvar/tl/SOILPARM.TBL
A      var/test/4dvar/tl/namelist.input.serial
A      var/test/4dvar/tl/GENPARM.TBL
A      var/test/4dvar/tl/LANDUSE.TBL
A      var/test/4dvar/tl/RRTM_DATA
A      var/test/4dvar/tl/trace
A      var/test/4dvar/radiance_info
A      var/test/4dvar/VEGPARM.TBL
A      var/test/4dvar/da_run_wrfplus_ad.ksh
A      var/test/4dvar/ad
A      var/test/4dvar/ad/VEGPARM.TBL
A      var/test/4dvar/ad/wrfbdy_d01
A      var/test/4dvar/ad/wrfinput_d01
A      var/test/4dvar/ad/namelist.input.parallel
A      var/test/4dvar/ad/SOILPARM.TBL
A      var/test/4dvar/ad/namelist.input.serial
A      var/test/4dvar/ad/GENPARM.TBL
A      var/test/4dvar/ad/LANDUSE.TBL
A      var/test/4dvar/ad/RRTM_DATA
A      var/test/4dvar/ad/trace
A      var/test/4dvar/README
A      var/test/4dvar/trace
A      var/test/4dvar/gmao_airs_bufr.tbl
A      var/test/4dvar/da_run_wrf_nl.ksh
A      var/test/4dvar/SOILPARM.TBL
A      var/test/4dvar/GENPARM.TBL
A      var/test/4dvar/nl
A      var/test/4dvar/nl/VEGPARM.TBL
A      var/test/4dvar/nl/wrfbdy_d01
A      var/test/4dvar/nl/wrfinput_d01
A      var/test/4dvar/nl/namelist.input.parallel
A      var/test/4dvar/nl/SOILPARM.TBL
A      var/test/4dvar/nl/namelist.input.serial
A      var/test/4dvar/nl/GENPARM.TBL
A      var/test/4dvar/nl/LANDUSE.TBL
A      var/test/4dvar/nl/RRTM_DATA
A      var/test/4dvar/namelist.input
A      var/test/4dvar/LANDUSE.TBL
A      var/test/4dvar/da_run_wrfplus_tl.ksh
A      var/test/4dvar/RRTM_DATA
A      var/test/tutorial
A  +   var/test/tutorial/namelist.input
A      var/test/update_bc
A  +   var/test/update_bc/parame.in
D      var/test/scripts
D      var/test/scripts/da_test_suite.ksh
D      var/test/scripts/da_test_etkf.ksh
D      var/test/scripts/da_compare_real.ksh
D      var/test/scripts/da_compare_region.ksh
D      var/test/scripts/da_compile_and_run.ksh
D      var/test/scripts/da_compare_suite.ksh
D      var/test/scripts/da_test_tests.ksh
D      var/test/scripts/da_compare_wrf.ksh
D      var/test/scripts/da_test_obs_tuning.ksh
D      var/test/scripts/da_test_defaults.ksh
D      var/test/scripts/da_compare_files.ksh
D      var/test/scripts/da_compare_netcdf.ksh
D      var/test/scripts/da_compare_wrfvar.ksh
D      var/test/scripts/da_test_gen_be.ksh
D      var/test/scripts/da_test_4dvar.ksh
D      var/test/scripts/da_compare_wps.ksh
D      var/test/scripts/da_compare_data.ksh
D      var/test/scripts/da_test_plotobs.ksh
D      var/test/scripts/da_test_verification.ksh
D      var/test/scripts/da_test_analysis.ksh
D      var/test/scripts/da_test_psot.ksh
D      var/test/scripts/da_compare_gen_be.ksh
D      var/test/namelist.3dvar_madis
D      var/test/RELEASE_NOTES.MADIS_WRF-Var
D      var/test/namelist.input
D      var/test/parame.in
D      var/test/README.MADIS_WRF-Var
A      var/test/madis
A  +   var/test/madis/namelist.3dvar_madis
A  +   var/test/madis/RELEASE_NOTES.MADIS_WRF-Var
A  +   var/test/madis/README.MADIS_WRF-Var
A  +   var/obsproc/namelist.obsproc.3dvar.wrfvar-tut
A      var/obsproc/namelist.obsproc.4dvar.wrfvar-tut
D      var/obsproc/namelist.obsproc.wrfvar-tut



git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3769 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ var/obsproc/MAP_plot/Map.csh
@@ -10,13 +10,13 @@
 #set echo
  unalias mv rm
 
- set Map_plot = /karri/users/guo/CODE/VAR/trunk/var/obsproc/MAP_plot
+ set Map_plot = /users/noname/WRFDA/var/obsproc/MAP_plot
 
- set TIME_WINDOW_MIN = '2007081221'
- set TIME_ANALYSIS   = '2007081300'
- set TIME_WINDOW_MAX = '2007081303'
+ set TIME_WINDOW_MIN = '2008020511'
+ set TIME_ANALYSIS   = '2008020512'
+ set TIME_WINDOW_MAX = '2008020513'
 
- set OBSDATA  = /karri3/guo/Tutorial/obs_gts_2007-08-13_00:00:00.3DVAR.GPS
+ set OBSDATA  = ../obs_gts_2008-02-05_12:00:00.3DVAR.GPS
 
 #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 #          ============== END OF CHANGES =================
@@ -78,9 +78,10 @@ cat >! namelist.file << EOF
   skewt_plot = .FALSE.,
  &
 EOF
- ln -sf ${Map_plot}/no_comment.sed    no_comment.sed
- ln -sf ${Map_plot}/mapco.tbl         mapco.tbl
- ln -sf ${Map_plot}/Map.exe           Map.exe
+
+if (! -e no_comment.sed ) ln -sf ${Map_plot}/no_comment.sed    no_comment.sed
+if (! -e mapco.tbl )      ln -sf ${Map_plot}/mapco.tbl         mapco.tbl
+if (! -e Map.exe )        ln -sf ${Map_plot}/Map.exe           Map.exe
 
  sed -f no_comment.sed  namelist.file >! tmp.file
  mv tmp.file  namelist.file
